26-FE: Fix router service error caused by circular reference
