+++ Okane.Client/src/shared/pages/RegisterPage.vue
@@ -7,13 +7,12 @@ import AuthForm from '@/features/auth/AuthForm.vue'
 import Heading from '@/shared/components/Heading.vue'
 import PageLayout from '@/shared/layouts/PageLayout.vue'
 
-import { ROUTE_NAME } from '@/shared/services/router/router.constants'
-
 import type { AuthFormState } from '@/features/auth/auth.types'
 
 import { useAuthStore } from '@/features/auth/useAuthStore'
 
 import { omitObjectKeys } from '@/shared/utils/object.utils'
+import { ROUTE_NAME } from '@/shared/services/router/router.service'
 
 const authStore = useAuthStore()
 const router = useRouter()
